refactored directory structure of sub-modules
