+++ ai4water/hyperopt/__init__.py
@@ -3,4 +3,4 @@
 from ._space import Real
 from ._space import Integer
 from ._space import Categorical
-from .hyper_opt import HyperOpt
+from ._main import HyperOpt
