+++ ai4water/postprocessing/SeqMetrics/__init__.py
@@ -2,7 +2,7 @@
 The values in a sequence are not necessarily related.
 """
 
-from ._SeqMetrics import Metrics
+from ._main import Metrics
 from ._regression import RegressionMetrics
 from ._classification import ClassificationMetrics
 from .utils import plot_metrics
