+++ ai4water/postprocessing/interpret/__init__.py
@@ -1,2 +1,2 @@
 
-from ._interpret import Interpret
+from ._main import Interpret
